% Auto-generated: do not edit by hand
\name{tableComponent}

\alias{tableComponent}

\title{TableComponent component}

\description{

}

\usage{
tableComponent(id=NULL, ascending=NULL, column=NULL, columns=NULL,
data=NULL, label=NULL, location=NULL, page_current=NULL,
page_size=NULL, report_number=NULL, total_results=NULL,
value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{ascending}{Numeric. }

\item{column}{Character. }

\item{columns}{Unnamed list. }

\item{data}{Unnamed list. }

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{location}{Character. }

\item{page_current}{Numeric. }

\item{page_size}{Numeric. }

\item{report_number}{Character. }

\item{total_results}{Numeric. }

\item{value}{Character. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

